Rename one-word components to multi-word ones
