+++ src/views/AppModal.vue
@@ -90,7 +90,7 @@
 
 <script setup lang="ts">
 import { ref } from 'vue'
-import Breadcrumb from '../partials/Breadcrumb.vue'
+import Breadcrumb from '../partials/AppBreadcrumb.vue'
 
 const open = ref(false)
 </script>
