+++ src/views/BlankView.vue
@@ -3,5 +3,5 @@
   <Breadcrumb breadcrumb="Blank" />
 </template>
 <script lang="ts" setup>
-import Breadcrumb from '../partials/Breadcrumb.vue'
+import Breadcrumb from '../partials/AppBreadcrumb.vue'
 </script>
