+++ src/views/CardView.vue
@@ -88,5 +88,5 @@
   </div>
 </template>
 <script lang="ts" setup>
-import Breadcrumb from '../partials/Breadcrumb.vue'
+import Breadcrumb from '../partials/AppBreadcrumb.vue'
 </script>
