+++ src/views/ChartView.vue
@@ -29,7 +29,7 @@
 </template>
 
 <script setup lang="ts">
-import Breadcrumb from "../partials/Breadcrumb.vue";
+import Breadcrumb from "../partials/AppBreadcrumb.vue";
 import DonutChart from "@/components/charts/DonutChart.vue";
 import LineChart from "@/components/charts/LineChart.vue";
 import BarChart from "../components/charts/BarChart.vue";
